Move discovery in a new module
